back/iranchat/requirements.txt
asgiref==3.5.0
Django==3.2.7
django-taggit==2.1.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
django-ckeditor==6.2.0
